+++ src/styles/index.scss
@@ -9,8 +9,8 @@ body {
 }
 
 .notification {
-  transition-property: opacity, min-height;
   transition-duration: 1s;
+  transition-property: opacity, min-height;
   min-height: 36px;
 }
 
@@ -20,5 +20,6 @@ body {
   pointer-events: none;
 }
 
+@import "./todo.scss";
 @import "./todoapp";
 @import "./filter";
